add callbacks

+++ model.py
@@ -1,3 +1,4 @@
+from datetime import datetime
 import matplotlib.pyplot as plt
 import pandas as pd
 import numpy as np
@@ -27,6 +28,8 @@ from utils import *
 from network import *
 from generator import *
 
+
+
 data_dir = 'data'
 
 print("Load Data...")
@@ -51,14 +54,21 @@ gen_train = generator(data_train, augment=True)
 gen_val = generator(data_val)
 
 log_dir = './logs'
+model_dir = './models'
+date_string = datetime.utcnow().strftime('%Y_%m_%d_%H_%M_%S_%f')[:-3]
 nb_epoch = 20
 nb_samples_per_epoch = 8000
 nb_val_samples = len(data_val)
 learning_rate = 1e-4
 min_delta=0.00001
 
-print('\nRemove logs')
-shutil.rmtree(log_dir)
+if os.path.exists(log_dir):
+	print('\nRemove logs')
+	shutil.rmtree(log_dir)
+
+if not os.path.exists(model_dir):
+	print('\nCreate model_dir')
+	os.mkdir(model_dir) 
 
 print("\nTrain Data")
 
@@ -68,18 +78,9 @@ history_object = model.fit_generator(gen_train,
                     validation_steps=nb_val_samples, 
                     epochs=nb_epoch,
                     callbacks=[
-                        keras.callbacks.EarlyStopping(
-                            monitor='val_loss', 
-                            min_delta=min_delta, 
-                            patience=1, 
-                            verbose=0, 
-                            mode='auto'
-                        ),
-                        keras.callbacks.TensorBoard(
-                            log_dir=log_dir, 
-                            histogram_freq=0, 
-                            write_graph=True, 
-                            write_images=True)
+				   keras.callbacks.ModelCheckpoint(model_dir + '/model.' + date_string + '.{epoch:02d}-{val_loss:.2f}.h5', monitor='val_loss', verbose=0, save_best_only=False, save_weights_only=False, mode='auto', period=1),
+                       keras.callbacks.EarlyStopping(monitor='val_loss', min_delta=min_delta, patience=1, verbose=0, mode='auto'),
+                       keras.callbacks.TensorBoard(log_dir=log_dir, histogram_freq=0, write_graph=True, write_images=True)
                     ])
